refactor: split shell host controller ownership
